requirements.txt
uvicorn
ctransformers
langchain
python-box
streamlit 
sentence-transformers
